Allow reading/writing emd NCEM file by splitting emd module into submodules for ncem and velox file format
